+++ jest.config.ts
@@ -11,6 +11,7 @@ const config: Config = {
   ],
   rootDir: 'src',
   testEnvironment: 'node',
+  setupFiles: ['dotenv/config'],
   setupFilesAfterEnv: ['jest-extended/all'],
   testMatch: ['**/*.spec.ts'],
   extensionsToTreatAsEsm: ['.ts'],