+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './firebase/index.js';
+export * from './identity-platform/index.js';
 export * from './validation/index.js';